Add transfer_scatter lowering
