+++ compiler/src/iree/compiler/Codegen/Dialect/VectorExt/Transforms/BUILD.bazel
@@ -38,7 +38,7 @@ iree_compiler_cc_library(
         "BufferizationInterfaces.cpp",
         "DistributionPatterns.cpp",
         "LowerTransferGatherScatterOps.cpp",
-        "LowerTransferGatherToGather.cpp",
+        "LowerTransferGatherScatterToVector.cpp",
         "Passes.cpp",
         "VectorExtFoldUnitExtentDims.cpp",
     ],
