+++ compiler/src/iree/compiler/Codegen/Dialect/VectorExt/Transforms/CMakeLists.txt
@@ -32,7 +32,7 @@ iree_cc_library(
     "BufferizationInterfaces.cpp"
     "DistributionPatterns.cpp"
     "LowerTransferGatherScatterOps.cpp"
-    "LowerTransferGatherToGather.cpp"
+    "LowerTransferGatherScatterToVector.cpp"
     "Passes.cpp"
     "VectorExtFoldUnitExtentDims.cpp"
   DEPS
