+++ compiler/src/iree/compiler/Codegen/Dialect/VectorExt/Transforms/test/CMakeLists.txt
@@ -14,7 +14,7 @@ iree_lit_test_suite(
   NAME
     lit
   SRCS
-    "lower_transfer_gather_to_vector_gather.mlir"
+    "lower_transfer_gather_scatter_to_vector.mlir"
     "vector_ext_fold_unit_extent_dims.mlir"
     "vectorize_vector_ext_ops.mlir"
   TOOLS
